Refactored patient compartment logic into separate file
