+++ packages/server/src/fhir/repo.ts
@@ -1,5 +1,4 @@
-import { accessDenied, AccessPolicy, allOk, assertOk, badRequest, Bundle, CompartmentDefinition, CompartmentDefinitionResource, created, deepEquals, Filter, getSearchParameterDetails, gone, isGone, isNotFound, isOk, Login, Meta, notFound, notModified, OperationOutcome, Operator as FhirOperator, parseFhirPath, Reference, Resource, SearchParameter, SearchParameterDetails, SearchRequest, SortRule, stringify } from '@medplum/core';
-import { readJson } from '@medplum/definitions';
+import { accessDenied, AccessPolicy, allOk, assertOk, badRequest, Bundle, created, deepEquals, Filter, getSearchParameterDetails, gone, isGone, isNotFound, isOk, Login, Meta, notFound, notModified, OperationOutcome, Operator as FhirOperator, parseFhirPath, Reference, Resource, SearchParameter, SearchParameterDetails, SearchRequest, SortRule, stringify } from '@medplum/core';
 import { randomUUID } from 'crypto';
 import { applyPatch, Operation } from 'fast-json-patch';
 import validator from 'validator';
@@ -8,6 +7,7 @@ import { MEDPLUM_PROJECT_ID, PUBLIC_PROJECT_ID } from '../constants';
 import { getClient } from '../database';
 import { addSubscriptionJobs } from '../workers/subscription';
 import { AddressTable, ContactPointTable, HumanNameTable, IdentifierTable, LookupTable } from './lookups';
+import { getPatientId, getPatientCompartmentResourceTypes } from './patient';
 import { validateResource, validateResourceType } from './schema';
 import { getSearchParameter, getSearchParameters } from './search';
 import { InsertQuery, Operator, SelectQuery } from './sql';
@@ -47,12 +47,6 @@ export interface RepositoryContext {
 }
 
 export type RepositoryResult<T extends Resource | undefined> = Promise<[OperationOutcome, T | undefined]>;
-
-/**
- * Patient compartment definitions.
- * See: https://www.hl7.org/fhir/compartmentdefinition-patient.html
- */
-let patientCompartment: CompartmentDefinition | undefined = undefined;
 
 /**
  * Public resource types are in the "public" project.
@@ -598,7 +592,7 @@ export class Repository {
       result.add(resolveId(resource.meta.account) as string);
     }
 
-    const patientId = getPatientCompartmentId(resource);
+    const patientId = getPatientId(resource);
     if (patientId) {
       result.add(patientId);
     }
@@ -763,7 +757,7 @@ export class Repository {
     }
 
     if (updated.resourceType !== 'Patient') {
-      const patientId = getPatientCompartmentId(updated);
+      const patientId = getPatientId(updated);
       if (patientId) {
         // If the resource is in a patient compartment, then lookup the patient.
         const [patientOutcome, patient] = await repo.readResource('Patient', patientId);
@@ -861,136 +855,6 @@ export class Repository {
     const { adminClientId } = getConfig();
     return !!adminClientId && this.context.author.reference === 'ClientApplication/' + adminClientId;
   }
-}
-
-/**
- * Returns the list of patient compartment properties, if the resource type is in a patient compartment.
- * Returns undefined otherwise.
- * See: https://www.hl7.org/fhir/compartmentdefinition-patient.html
- * @param resourceType The resource type.
- * @returns List of property names if in patient compartment; undefined otherwise.
- */
-function getPatientCompartmentProperties(resourceType: string): string[] | undefined {
-  const resourceList = getPatientCompartments().resource as CompartmentDefinitionResource[];
-  for (const resource of resourceList) {
-    if (resource.code === resourceType) {
-      return resource.param;
-    }
-  }
-  return undefined;
-}
-
-/**
- * Returns the list of patient resource types.
- * See: https://www.hl7.org/fhir/compartmentdefinition-patient.html
- * @returns List of resource types in the patient compartment.
- */
-function getPatientCompartmentResourceTypes(): string[] {
-  const result = ['Patient'];
-  const resourceList = getPatientCompartments().resource as CompartmentDefinitionResource[];
-  for (const resource of resourceList) {
-    if (resource.code && resource.param) {
-      // Only add resource definitions with a 'param' value
-      // The param value defines the eligible properties
-      // If param is missing, it means the resource type is not in the compartment
-      result.push(resource.code);
-    }
-  }
-  return result;
-}
-
-function getPatientCompartments(): CompartmentDefinition {
-  if (!patientCompartment) {
-    patientCompartment = readJson('fhir/r4/compartmentdefinition-patient.json') as CompartmentDefinition;
-  }
-  return patientCompartment;
-}
-
-/**
- * Returns the patient compartment ID for a resource.
- * If the resource is in a patient compartment (i.e., an Observation about the patient),
- * then return the patient ID.
- * If the resource is not in a patient compartment (i.e., a StructureDefinition),
- * then return undefined.
- * @param resource The resource to inspect.
- * @returns The patient ID if found; undefined otherwise.
- */
-function getPatientCompartmentId(resource: Resource): string | undefined {
-  if (resource.resourceType === 'Patient') {
-    return resource.id;
-  }
-  const patientCompartmentProperties = getPatientCompartmentProperties(resource.resourceType);
-  if (patientCompartmentProperties) {
-    return getPatientId(resource, patientCompartmentProperties);
-  }
-  return undefined;
-}
-
-/**
- * Returns the patient ID from a resource.
- * See Patient Compatment: https://www.hl7.org/fhir/compartmentdefinition-patient.json.html
- * @param resource The resource.
- * @returns The patient ID if found; undefined otherwise.
- */
-export function getPatientId(resource: Resource, properties: string[]): string | undefined {
-  if (resource.resourceType === 'Patient') {
-    return resource.id;
-  }
-
-  for (const property of properties) {
-    if (property in resource) {
-      const value: Reference | Reference[] | undefined = (resource as any)[property];
-      const patientId = getPatientIdFromReferenceProperty(value);
-      if (patientId) {
-        return patientId;
-      }
-    }
-  }
-
-  return undefined;
-}
-
-/**
- * Tries to return a patient ID from a reference or array of references.
- * @param reference A FHIR reference or array of references.
- * @returns The patient ID if found; undefined otherwise.
- */
-function getPatientIdFromReferenceProperty(reference: Reference | Reference[] | undefined): string | undefined {
-  if (!reference) {
-    return undefined;
-  }
-  if (Array.isArray(reference)) {
-    return getPatientIdFromReferenceArray(reference);
-  } else {
-    return getPatientIdFromReference(reference);
-  }
-}
-
-/**
- * Tries to return a patient ID from an array of references.
- * @param references Array of FHIR references.
- * @returns The patient ID if found; undefined otherwise.
- */
-function getPatientIdFromReferenceArray(references: Reference[]): string | undefined {
-  for (const reference of references) {
-    const result = getPatientIdFromReference(reference);
-    if (result) {
-      return result;
-    }
-  }
-  return undefined;
-}
-
-/**
- * Tries to return a patient ID from a FHIR reference.
- * @param reference A FHIR reference.
- * @returns The patient ID if found; undefined otherwise.
- */
-function getPatientIdFromReference(reference: Reference): string | undefined {
-  if (reference.reference?.startsWith('Patient/')) {
-    return resolveId(reference);
-  }
-  return undefined;
 }
 
 /**
